Use dedicated types for all Matter IDs

+++ app/src/main/java/io/aether/android/chip/SubscriptionHelper.kt
@@ -11,6 +11,7 @@ import chip.devicecontroller.model.ChipAttributePath
 import chip.devicecontroller.model.ChipEventPath
 import chip.devicecontroller.model.ChipPathId
 import chip.devicecontroller.model.NodeState
+import io.aether.android.matter.toLong
 import java.lang.Exception
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -79,9 +80,9 @@ class SubscriptionHelper @Inject constructor(private val chipClient: ChipClient)
     nodeState.endpointStates.forEach { (_endpointId, endpointState) ->
       if (_endpointId != endpointId) return@forEach
       endpointState.clusterStates.forEach { (clusterId, clusterState) ->
-        if (clusterId != clusterAttribute.clusterId) return@forEach
+        if (clusterId != clusterAttribute.clusterId.toLong()) return@forEach
         clusterState.attributeStates.forEach { (attributeId, attributeState) ->
-          if (attributeId != clusterAttribute.attributeId) return@forEach
+          if (attributeId != clusterAttribute.attributeId.toLong()) return@forEach
           return attributeState.value
         }
       }
